Add measure data from hynix

+++ models/nipq_quantization_module.py
@@ -70,9 +70,9 @@ class Quantizer(nn.Module):
     def hwnoise_init(self, cbits, mapping_mode, co_noise=0.01, noise_type='prop', res_val='rel', max_epoch=-1):
         bit = 2 + torch.sigmoid(self.bit)*12
         w_format = 'state' if res_val == 'abs' or noise_type == 'meas' else 'weight'
-        self.noise_cell = Noise_cell(bit.round().squeeze(), cbits, mapping_mode, co_noise, noise_type=noise_type, \
+        self.noise_cell = Noise_cell(int(bit.round().squeeze()), cbits, mapping_mode, co_noise, noise_type=noise_type, \
                                     res_val=res_val, w_format=w_format, max_epoch=max_epoch)
-        self.inf_noise_cell = Noise_cell(bit.round().squeeze(), cbits, mapping_mode, co_noise, noise_type='prop', \
+        self.inf_noise_cell = Noise_cell(int(bit.round().squeeze()), cbits, mapping_mode, co_noise, noise_type='prop', \
                                     res_val=res_val, w_format=w_format, max_epoch=max_epoch)
     
     def get_alpha(self):
